Moved worker functions for db to utils/db.rs.
