+++ src/utils/mod.rs
@@ -4,6 +4,7 @@ pub mod streams;
 pub mod command;
 pub mod sequence;
 pub mod defs;
+pub mod db;
 
 use clap::{Parser, ValueEnum};
